Fix record validation
Record validation is too permissive if one just checks to see if all
of the fields in a candidate schema validate against a record.  The
record may have additional fields which are not validated, and those
fields will ultimately get dropped.  Furthermore, if the two record
types are part of a union and one has a subset of fields that the
other has, the record type could be serialized incorrectly.